oracle-ai-database-gcp-vertex-ai/rag-ai-agent-python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
oracledb==2.0.0
google-cloud-aiplatform==1.70.0
vertexai==1.0.0
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-google-vertexai==2.0.0
gunicorn==21.2.0
